requirements.txt
Django>=4.2.0,<5.0.0
Pillow>=10.0.0
django-modeltranslation>=0.18.12

